function_app/requirements.txt
pocketflow>=0.0.1
pyyaml>=6.0
requests>=2.28.0
gitpython>=3.1.0
google-cloud-aiplatform>=1.25.0
google-genai>=1.9.0
python-dotenv>=1.0.0
Flask
flask-cors
azure-storage-blob==12.25.1
azure-functions==1.23.0
azure-storage-queue==12.12.0